requirements.txt
streamlit
openai
supabase
python-dotenv
google-generativeai
Pillow
watchdog